Play Solo

This function demonstrates a solo player chute-and-ladders game with a user defined board. You can also: Define the board with define_board() function. Visualize the board with show_board() function. Perform simulation with simualte_board() function.

play_solo(theboard, verbose = FALSE)

Arguments

theboard

A list of four elements containing board information:height,width,ladderand chute

verbose

A logical T/F indicating whether to show the function output for each turn

Value

if verbose=F, return a list of play log: turns:the number of total turns, ladder_tally: the frequency of use of each ladder; chute_tally:the frequency of use of each ladder; and move_log: record of all spaces that the user landed if verbose=T, return a list of the above play log and displays each move of this play: The start of each turn begins with the turn number: Turn x

It announces where the player is located at the start of the turn: Start at x

It announces what the player spun: Spinner: x

If the player lands on a special space, it says Landed on: x

It will announce Chute! or Ladder! depending on what type of special space it is. It announces where the player ends the turn: Turn ends at: x

Examples

board <- define_board(height=10, width=10,
ladder=list(c(1, 23), c(9,36), c(50,75), c(68,98)),
chute=list(c(26,8), c(78,48), c(35,6), c(90,69)))

play_solo(board, verbose = TRUE)
#> Turn  1 
#> Start at  0 
#> Spinner  5 
#> Landed on  5 
#> 
#> Turn  2 
#> Start at  5 
#> Spinner  4 
#> Landed on  9 
#> Ladder! 
#> Turn ends at  36 
#> 
#> Turn  3 
#> Start at  36 
#> Spinner  5 
#> Landed on  41 
#> 
#> Turn  4 
#> Start at  41 
#> Spinner  6 
#> Landed on  47 
#> 
#> Turn  5 
#> Start at  47 
#> Spinner  4 
#> Landed on  51 
#> 
#> Turn  6 
#> Start at  51 
#> Spinner  6 
#> Landed on  57 
#> 
#> Turn  7 
#> Start at  57 
#> Spinner  1 
#> Landed on  58 
#> 
#> Turn  8 
#> Start at  58 
#> Spinner  5 
#> Landed on  63 
#> 
#> Turn  9 
#> Start at  63 
#> Spinner  5 
#> Landed on  68 
#> Ladder! 
#> Turn ends at  98 
#> 
#> Turn  10 
#> Start at  98 
#> Spinner  5 
#> Landed on  103 
#> 
#> Turn  12 
#> Start at  99 
#> Spinner  2 
#> Landed on  101 
#> 
#> Turn  14 
#> Start at  99 
#> Spinner  6 
#> Landed on  105 
#> 
#> Turn  16 
#> Start at  99 
#> Spinner  3 
#> Landed on  102 
#> 
#> Turn  18 
#> Start at  99 
#> Spinner  6 
#> Landed on  105 
#> 
#> Turn  20 
#> Start at  99 
#> Spinner  3 
#> Landed on  102 
#> 
#> Turn  22 
#> Start at  99 
#> Spinner  5 
#> Landed on  104 
#> 
#> Turn  24 
#> Start at  99 
#> Spinner  5 
#> Landed on  104 
#> 
#> Turn  26 
#> Start at  99 
#> Spinner  4 
#> Landed on  103 
#> 
#> Turn  28 
#> Start at  99 
#> Spinner  5 
#> Landed on  104 
#> 
#> Turn  30 
#> Start at  99 
#> Spinner  2 
#> Landed on  101 
#> 
#> Turn  32 
#> Start at  99 
#> Spinner  6 
#> Landed on  105 
#> 
#> Turn  34 
#> Start at  99 
#> Spinner  4 
#> Landed on  103 
#> 
#> Turn  36 
#> Start at  99 
#> Spinner  2 
#> Landed on  101 
#> 
#> Turn  38 
#> Start at  99 
#> Spinner  3 
#> Landed on  102 
#> 
#> Turn  40 
#> Start at  99 
#> Spinner  1 
#> Landed on  100 
#> 
#> $turns
#> [1] 40
#> 
#> $ladder_tally
#> [1] 0 1 0 1
#> 
#> $chute_tally
#> [1] 0 0 0 0
#> 
#> $move_log
#>  [1]   0   5  36  41  47  51  57  58  63  98 103  99 101  99 105  99 102  99 105
#> [20]  99 102  99 104  99 104  99 103  99 104  99 101  99 105  99 103  99 101  99
#> [39] 102  99 100
#>